refactor(renderer): change file name
